remove backup from runway
