+++ consensus/src/lib.rs
@@ -19,8 +19,8 @@ pub use aleph_bft_types::{
     PartialMultisignature, PartiallyMultisigned, Recipient, Round, SessionId, Signable, Signature,
     SignatureError, SignatureSet, Signed, SpawnHandle, TaskHandle, UncheckedSigned,
 };
-pub use config::{default_config, exponential_slowdown, Config, DelayConfig, LocalIO};
-pub use member::run_session;
+pub use config::{default_config, exponential_slowdown, Config, DelayConfig};
+pub use member::{run_session, LocalIO};
 pub use network::NetworkData;
 
 #[cfg(test)]
